requirements.txt
# ADME-Tox Predictor - Dependencies
# AI-Based Drug Safety & Toxicity Prediction System

# Web Framework
streamlit==1.31.0

# Scientific Computing
numpy>=1.24.0
pandas==2.1.4

# Machine Learning (Fixed versions for model compatibility)
scikit-learn==1.5.1
joblib==1.4.2

# Chemistry & Molecular Descriptors
rdkit>=2023.9.0
mordred==1.2.0

# Visualization
plotly==5.18.0

# Utilities
Pillow>=10.0.0
